requirements.txt
# ============================================
# ASKLY - Requirements
# Clean Architecture Flask Application
# ============================================

# -------- Core Framework --------
Flask>=3.0,<4.0
Werkzeug>=3.0

# -------- Database --------
Flask-SQLAlchemy>=3.1
SQLAlchemy>=2.0
psycopg2-binary>=2.9

# -------- Migrations --------
Flask-Migrate>=4.0
alembic>=1.13

# -------- Authentication --------
Flask-Login>=0.6
PyJWT>=2.8

# -------- API & Documentation --------
flask-restx>=1.3
flask-cors>=4.0

# -------- Forms & Validation --------
Flask-WTF>=1.2
WTForms>=3.1
email-validator>=2.0
pydantic>=2.5

# -------- Security --------
Flask-Talisman>=1.1
Flask-Limiter>=3.5
bcrypt>=4.1

# -------- Environment --------
python-dotenv>=1.0

# -------- Utilities --------
python-dateutil>=2.8

# -------- Testing --------
pytest>=8.0
pytest-flask>=1.3
pytest-cov>=4.1
pytest-mock>=3.12
factory-boy>=3.3
faker>=22.0
httpx>=0.26

# -------- Code Quality --------
black>=24.0
isort>=5.13
flake8>=7.0
mypy>=1.8
pre-commit>=3.6

# -------- Type Stubs --------
types-python-dateutil>=2.8
